2024-07-02 ver 2.3.0-dev - TM5887 LiDAR SDK Update

+++ libhesai/PtcClient/include/tcp_client.h
@@ -40,7 +40,7 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <winsock2.h>
 #include <ws2tcpip.h> 
 #else
-typedef unsigned int SOCKET;
+typedef signed long SOCKET;
 #endif
 #include "client_base.h"
 #include <stdint.h>
